GitFlowService

+++ src/main/java/com/feeltens/git/service/impl/GitFlowServiceImpl.java
@@ -2,6 +2,7 @@ package com.feeltens.git.service.impl;
 
 import cn.hutool.core.collection.CollUtil;
 import cn.hutool.core.date.DateUtil;
+import cn.hutool.core.util.ObjUtil;
 import cn.hutool.core.util.StrUtil;
 import com.alibaba.fastjson.JSON;
 import com.alibaba.fastjson.JSONObject;
@@ -63,6 +64,7 @@ import com.feeltens.git.vo.req.AddIntoMixBranchReqVO;
 import com.feeltens.git.vo.req.CreateGitBranchReqVO;
 import com.feeltens.git.vo.req.DeleteGitBranchReqVO;
 import com.feeltens.git.vo.req.ListGitBranchReqVO;
+import com.feeltens.git.vo.req.ListGitRepositoryNameReqVO;
 import com.feeltens.git.vo.req.PageGitBranchReqVO;
 import com.feeltens.git.vo.req.PageGitOrganizationReqVO;
 import com.feeltens.git.vo.req.PageGitProjectReqVO;
@@ -252,9 +254,9 @@ public class GitFlowServiceImpl implements GitFlowService {
         gitOrganizationDO.setOrganizationName(organizationsVO.getOrganizationName());
         gitOrganizationDO.setDescription(organizationsVO.getDescription());
         gitOrganizationDO.setWebUrl(null);
-        gitOrganizationDO.setCreateBy(GitConstant.SYSTEM_NAME);
+        gitOrganizationDO.setCreateBy(req.getOperator());
         gitOrganizationDO.setCreateTime(organizationsVO.getCreateTime());
-        gitOrganizationDO.setUpdateBy(GitConstant.SYSTEM_NAME);
+        gitOrganizationDO.setUpdateBy(req.getOperator());
         gitOrganizationDO.setUpdateTime(organizationsVO.getUpdateTime());
         gitOrganizationDO.setDeleted(0L);
 
@@ -292,10 +294,14 @@ public class GitFlowServiceImpl implements GitFlowService {
                 throw new RuntimeException("入参为空");
             }
             req.setOrganizationId(StrUtil.trim(req.getOrganizationId()));
+            req.setRepositoryName(StrUtil.trim(req.getRepositoryName()));
             req.setProjectName(StrUtil.trim(req.getProjectName()));
             req.setOperator(StrUtil.trim(req.getOperator()));
             if (StrUtil.isEmptyIfStr(req.getOrganizationId())) {
                 throw new RuntimeException("组织id不能为空");
+            }
+            if (StrUtil.isEmptyIfStr(req.getRepositoryName())) {
+                throw new RuntimeException("请选择仓库");
             }
             if (StrUtil.isEmptyIfStr(req.getProjectName())) {
                 throw new RuntimeException("工程名称不能为空");
@@ -311,7 +317,7 @@ public class GitFlowServiceImpl implements GitFlowService {
             }
 
             // 调用open api，查询git仓库信息
-            ListRepositoriesReq request = new ListRepositoriesReq(req.getProjectName()); // 仓库名称
+            ListRepositoriesReq request = new ListRepositoriesReq(req.getRepositoryName()); // 仓库名称
             setGitMergeFlowConfig(request, req.getOrganizationId());
 
             ListRepositoriesResp oapiResponse = gitOpenApiFactory.listRepositories(request);
@@ -404,9 +410,9 @@ public class GitFlowServiceImpl implements GitFlowService {
         gitMixBranchDTO.setEnv(envEnum.getCode());
         // gitMixBranchDTO.setProjectId(); // set later
         gitMixBranchDTO.setAllMergeFlag(1);
-        gitMixBranchDTO.setCreateBy(GitConstant.SYSTEM_NAME);
+        gitMixBranchDTO.setCreateBy(operator);
         gitMixBranchDTO.setCreateTime(now);
-        gitMixBranchDTO.setUpdateBy(GitConstant.SYSTEM_NAME);
+        gitMixBranchDTO.setUpdateBy(operator);
         gitMixBranchDTO.setUpdateTime(now);
         gitMixBranchDTO.setDeleted(0L);
         gitMixBranchDTO.setMixBranchExt(null);
@@ -549,6 +555,21 @@ public class GitFlowServiceImpl implements GitFlowService {
     }
 
     @Override
+    public List<String> listGitRepositoryNameByOpenApi(ListGitRepositoryNameReqVO req) {
+        // 调用open api，查询git仓库信息
+        ListRepositoriesReq request = new ListRepositoriesReq(); // 仓库名称
+        setGitMergeFlowConfig(request, req.getOrganizationId());
+
+        ListRepositoriesResp oapiResponse = gitOpenApiFactory.listRepositories(request);
+        if (null == oapiResponse || CollUtil.isEmpty(oapiResponse.getRepoItemList())) {
+            throw new BizException("未找到仓库信息 organizationId:" + req.getOrganizationId());
+        }
+
+        return oapiResponse.getRepoItemList().stream()
+                .map(ListRepositoriesResp.ListRepositoriesRespItem::getName).collect(Collectors.toList());
+    }
+
+    @Override
     public List<PageGitProjectRespVO> listGitProject() {
         List<GitProjectDO> list = gitProjectMapper.pageProject(new PageGitProjectReqVO(), null, null);
         return GitProjectConverter.toPageVos(list);
@@ -645,10 +666,6 @@ public class GitFlowServiceImpl implements GitFlowService {
 
         List<GitBranchDO> list = Lists.newArrayList();
         for (GitBranchDO gitBranchDO : branchList) {
-            // 默认分支，则跳过
-            if (gitProjectDb.getDefaultBranch().equals(gitBranchDO.getBranchName())) {
-                continue;
-            }
             // 中间分支，则跳过
             if (GitConstant.EXCLUDE_BRANCH_NAME_LIST.contains(gitBranchDO.getBranchName())) {
                 continue;
@@ -834,7 +851,7 @@ public class GitFlowServiceImpl implements GitFlowService {
         // 2、合并主分支到中间分支
         DoMergeBranchResultDTO mergeResultDTO = doMergeIntoMixBranch(
                 req.getMixBranchId(), gitProjectDb.getRepositoryId(), gitProjectDb.getOrganizationId(),
-                gitProjectDb.getDefaultBranch(), mixBranchDb.getMixBranchName());
+                gitProjectDb.getDefaultBranch(), mixBranchDb.getMixBranchName(), req.getOperator());
         if (!mergeResultDTO.getFlag()) {
             int updateRow = gitMixBranchMapper.updateAllMergeFlag(req.getMixBranchId(), 0, req.getOperator());
             // 自动合并失败，则返回手动集成shell
@@ -846,7 +863,8 @@ public class GitFlowServiceImpl implements GitFlowService {
             for (String branchName : mixedBranchNameList) {
                 // 合并源分支到目标分支
                 DoMergeBranchResultDTO mergeBranchResultDTO = doMergeIntoMixBranch(
-                        req.getMixBranchId(), gitProjectDb.getRepositoryId(), gitProjectDb.getOrganizationId(), branchName, mixBranchDb.getMixBranchName());
+                        req.getMixBranchId(), gitProjectDb.getRepositoryId(), gitProjectDb.getOrganizationId(), branchName,
+                        mixBranchDb.getMixBranchName(), req.getOperator());
                 if (!mergeBranchResultDTO.getFlag()) {
                     int updateRow = gitMixBranchMapper.updateAllMergeFlag(req.getMixBranchId(), 0, req.getOperator());
                     // 自动合并失败，则返回手动集成shell
@@ -927,7 +945,7 @@ public class GitFlowServiceImpl implements GitFlowService {
         for (String branchName : needMixBranchNameList) {
             // 合并源分支到目标分支
             DoMergeBranchResultDTO mergeBranchResultDTO = doMergeIntoMixBranch(
-                    req.getMixBranchId(), gitProjectDb.getRepositoryId(), gitProjectDb.getOrganizationId(), branchName, mixBranchDb.getMixBranchName());
+                    req.getMixBranchId(), gitProjectDb.getRepositoryId(), gitProjectDb.getOrganizationId(), branchName, mixBranchDb.getMixBranchName(), req.getOperator());
             if (!mergeBranchResultDTO.getFlag()) {
                 int updateRow = gitMixBranchMapper.updateAllMergeFlag(req.getMixBranchId(), 0, req.getOperator());
                 // 自动合并失败，则返回手动集成shell
@@ -1001,8 +1019,7 @@ public class GitFlowServiceImpl implements GitFlowService {
         if (null == req) {
             throw new BizException("参数错误");
         }
-        // req.setOperator(StrUtil.trim(req.getOperator()));
-        req.setOperator(GitConstant.SYSTEM_NAME);
+        req.setOperator(StrUtil.trim(req.getOperator()));
         if (StrUtil.isEmptyIfStr(req.getOperator())) {
             throw new BizException("操作人不能为空");
         }
@@ -1034,6 +1051,10 @@ public class GitFlowServiceImpl implements GitFlowService {
             }
         }
 
+        if (req.getGitBranchNameList().contains(gitProjectDb.getDefaultBranch())) {
+            throw new BizException("不能直接删除主分支，请联系管理员 " + gitProjectDb.getDefaultBranch());
+        }
+
         // 成功删除的分支名list
         List<String> successDelBranchNameList = Lists.newArrayList();
         for (String branchName : req.getGitBranchNameList()) {
@@ -1057,7 +1078,14 @@ public class GitFlowServiceImpl implements GitFlowService {
         }
 
         // 落库
-        gitBranchMapper.deleteByProjectIdAndBranchNameList(gitProjectDb.getProjectId(), successDelBranchNameList, req.getOperator());
+        if (CollUtil.isNotEmpty(successDelBranchNameList)) {
+            gitBranchMapper.deleteByProjectIdAndBranchNameList(gitProjectDb.getProjectId(), successDelBranchNameList, req.getOperator());
+        }
+
+        if (ObjUtil.notEqual(req.getGitBranchNameList().size(), successDelBranchNameList.size())) {
+            log.warn("deleteGitBranch successDelBranchNameList:{}", JSON.toJSONString(successDelBranchNameList));
+            throw new BizException("有分支清理失败");
+        }
 
         return new DeleteGitBranchRespVO(successDelBranchNameList);
     }
@@ -1144,7 +1172,7 @@ public class GitFlowServiceImpl implements GitFlowService {
             createGitBranchReqVO.setGitProjectId(gitProjectDb.getProjectId());
             createGitBranchReqVO.setBranchDesc("基于当前中间分支，创建备份分支");
             createGitBranchReqVO.setSourceBranch(mixBranchName);
-            createGitBranchReqVO.setOperator(GitConstant.SYSTEM_NAME);
+            createGitBranchReqVO.setOperator(operator);
             backupBranchDO = buildGitBranchDO(createGitBranchReqVO, createBackupBranchResp);
 
             // 删除旧中间分支
@@ -1291,13 +1319,15 @@ public class GitFlowServiceImpl implements GitFlowService {
      * @param repositoryId     仓库ID
      * @param sourceBranchName 源分支名称
      * @param targetBranchName 目标分支名称
+     * @param operator         操作人
      * @return
      */
     private DoMergeBranchResultDTO doMergeIntoMixBranch(Long mixBranchId,
                                                         Long repositoryId,
                                                         String organizationId,
                                                         String sourceBranchName,
-                                                        String targetBranchName) {
+                                                        String targetBranchName,
+                                                        String operator) {
         String titleTemplate = GitConstant.SYSTEM_NAME + ": merge branch %s into %s";
         String title = String.format(titleTemplate, sourceBranchName, targetBranchName);
 
@@ -1327,7 +1357,7 @@ public class GitFlowServiceImpl implements GitFlowService {
         if (Objects.equals(MergeTotalStatusEnum.MERGED.getStatus(), createChangeRequestResp.getMergeTotalStatus())) {
             // 已合并
             gitMixBranchMapper.updateMergeRequestId(mixBranchId, null);
-            gitMixBranchItemMapper.updateMergeFlag(mixBranchId, sourceBranchName, 1, GitConstant.SYSTEM_NAME);
+            gitMixBranchItemMapper.updateMergeFlag(mixBranchId, sourceBranchName, 1, operator);
             return new DoMergeBranchResultDTO(mixBranchId, true, null);
         }
 
@@ -1342,7 +1372,7 @@ public class GitFlowServiceImpl implements GitFlowService {
             if (Objects.equals(MergeTotalStatusEnum.MERGED.getStatus(), mergeChangeRequestResp.getMergeTotalStatus())) {
                 // 已合并
                 gitMixBranchMapper.updateMergeRequestId(mixBranchId, null);
-                gitMixBranchItemMapper.updateMergeFlag(mixBranchId, sourceBranchName, 1, GitConstant.SYSTEM_NAME);
+                gitMixBranchItemMapper.updateMergeFlag(mixBranchId, sourceBranchName, 1, operator);
                 return new DoMergeBranchResultDTO(mixBranchId, true, null);
             }
         }
@@ -1360,7 +1390,7 @@ public class GitFlowServiceImpl implements GitFlowService {
             if (Objects.equals(MergeTotalStatusEnum.MERGED.getStatus(), getChangeRequestResp.getMergeTotalStatus())) {
                 // 已合并
                 gitMixBranchMapper.updateMergeRequestId(mixBranchId, null);
-                gitMixBranchItemMapper.updateMergeFlag(mixBranchId, sourceBranchName, 1, GitConstant.SYSTEM_NAME);
+                gitMixBranchItemMapper.updateMergeFlag(mixBranchId, sourceBranchName, 1, operator);
                 return new DoMergeBranchResultDTO(mixBranchId, true, null);
             }
 
@@ -1393,14 +1423,14 @@ public class GitFlowServiceImpl implements GitFlowService {
             if (Objects.equals(MergeTotalStatusEnum.MERGED.getStatus(), mergeChangeRequestResp.getMergeTotalStatus())) {
                 // 已合并
                 gitMixBranchMapper.updateMergeRequestId(mixBranchId, null);
-                gitMixBranchItemMapper.updateMergeFlag(mixBranchId, sourceBranchName, 1, GitConstant.SYSTEM_NAME);
+                gitMixBranchItemMapper.updateMergeFlag(mixBranchId, sourceBranchName, 1, operator);
                 return new DoMergeBranchResultDTO(mixBranchId, true, null);
             }
         }
 
         // 需要用户手动处理冲突
         gitMixBranchMapper.updateMergeRequestId(mixBranchId, createChangeRequestResp.getLocalId()); // set （git服务平台的）合并请求的id
-        gitMixBranchItemMapper.updateMergeFlag(mixBranchId, sourceBranchName, 0, GitConstant.SYSTEM_NAME);
+        gitMixBranchItemMapper.updateMergeFlag(mixBranchId, sourceBranchName, 0, operator);
         return new DoMergeBranchResultDTO(mixBranchId, false, getUserMergeShell(sourceBranchName, targetBranchName));
     }
